Split the offer models module into separate files.
This makes the following (last) models abstract:
 - Benefit
 - Condition
 - ProductRangeProductFile

The benefit and condition implementations are moved to their own
modules (`benefits` and `conditions`). The offer results are moved to
the `results` module. The `utils.Applicator` is moved to the new
`applicator` module.

The API is currently still backwards compatible
