+++ src/oscar/apps/offer/models.py
@@ -1,60 +1,16 @@
-import os
-import re
-import operator
-from decimal import Decimal as D, ROUND_DOWN, ROUND_UP
-
-from django.core import exceptions
-from django.db import models
-from django.db.models.query import Q
-from django.utils.encoding import python_2_unicode_compatible
-from django.utils.timezone import now
-from django.utils.translation import ungettext, ugettext_lazy as _
-from django.utils import six
-from django.core.exceptions import ValidationError
-from django.core.urlresolvers import reverse
-from django.conf import settings
-
-from oscar.core.compat import AUTH_USER_MODEL
-from oscar.core.loading import get_class, get_model
-from oscar.templatetags.currency_filters import currency
-from oscar.models import fields
-
 from oscar.core.loading import is_model_registered
-from oscar.apps.offer.abstract_models import *  # noqa
-
-__all__ = ['Condition', 'Benefit', 'CountCondition',
-           'CoverageCondition', 'ValueCondition', 'PercentageDiscountBenefit',
-           'AbsoluteDiscountBenefit', 'FixedPriceBenefit', 'ShippingBenefit',
-           'MultibuyDiscountBenefit', 'ShippingAbsoluteDiscountBenefit',
-           'ShippingFixedPriceBenefit', 'ShippingPercentageDiscountBenefit',
-           'RangeProductFileUpload',
-
-           'ApplicationResult', 'BasketDiscount', 'ShippingDiscount',
-           'PostOrderAction',
-
-           'ZERO_DISCOUNT', 'SHIPPING_DISCOUNT']
+from oscar.apps.offer.results import (
+    BasketDiscount, ShippingDiscount, PostOrderAction,
+    SHIPPING_DISCOUNT, ZERO_DISCOUNT)
+from oscar.apps.offer.abstract_models import (
+    AbstractBenefit, AbstractCondition, AbstractConditionalOffer,
+    AbstractRange, AbstractRangeProduct, AbstractRangeProductFileUpload)
 
 
-def range_anchor(range):
-    return u'<a href="%s">%s</a>' % (
-        reverse('dashboard:range-update', kwargs={'pk': range.pk}),
-        range.name)
-
-
-def unit_price(offer, line):
-    """
-    Return the relevant price for a given basket line.
-
-    This is required so offers can apply in circumstances where tax isn't known
-    """
-    return line.unit_effective_price
-
-
-def apply_discount(line, discount, quantity):
-    """
-    Apply a given discount to the passed basket
-    """
-    line.discount(discount, quantity, incl_tax=False)
+__all__ = [
+    'BasketDiscount', 'ShippingDiscount', 'PostOrderAction',
+    'SHIPPING_DISCOUNT', 'ZERO_DISCOUNT'
+]
 
 
 if not is_model_registered('offer', 'ConditionalOffer'):
@@ -64,367 +20,18 @@ if not is_model_registered('offer', 'ConditionalOffer'):
     __all__.append('ConditionalOffer')
 
 
-@python_2_unicode_compatible
-class Condition(models.Model):
-    """
-    A condition for an offer to be applied. You can either specify a custom
-    proxy class, or need to specify a type, range and value.
-    """
-    COUNT, VALUE, COVERAGE = ("Count", "Value", "Coverage")
-    TYPE_CHOICES = (
-        (COUNT, _("Depends on number of items in basket that are in "
-                  "condition range")),
-        (VALUE, _("Depends on value of items in basket that are in "
-                  "condition range")),
-        (COVERAGE, _("Needs to contain a set number of DISTINCT items "
-                     "from the condition range")))
-    range = models.ForeignKey(
-        'offer.Range', verbose_name=_("Range"), null=True, blank=True)
-    type = models.CharField(_('Type'), max_length=128, choices=TYPE_CHOICES,
-                            blank=True)
-    value = fields.PositiveDecimalField(
-        _('Value'), decimal_places=2, max_digits=12, null=True, blank=True)
-
-    proxy_class = fields.NullCharField(
-        _("Custom class"), max_length=255, unique=True, default=None)
-
-    class Meta:
-        app_label = 'offer'
-        verbose_name = _("Condition")
-        verbose_name_plural = _("Conditions")
-
-    def proxy(self):
-        """
-        Return the proxy model
-        """
-        klassmap = {
-            self.COUNT: CountCondition,
-            self.VALUE: ValueCondition,
-            self.COVERAGE: CoverageCondition}
-        # Short-circuit logic if current class is already a proxy class.
-        if self.__class__ in klassmap.values():
-            return self
-
-        field_dict = dict(self.__dict__)
-        for field in list(field_dict.keys()):
-            if field.startswith('_'):
-                del field_dict[field]
-
-        if self.proxy_class:
-            klass = load_proxy(self.proxy_class)
-            # Short-circuit again.
-            if self.__class__ == klass:
-                return self
-            return klass(**field_dict)
-        if self.type in klassmap:
-            return klassmap[self.type](**field_dict)
-        raise RuntimeError("Unrecognised condition type (%s)" % self.type)
-
-    def __str__(self):
-        return self.name
-
-    @property
-    def name(self):
-        """
-        A plaintext description of the condition. Every proxy class has to
-        implement it.
-
-        This is used in the dropdowns within the offer dashboard.
-        """
-        return self.proxy().name
-
-    @property
-    def description(self):
-        """
-        A description of the condition.
-        Defaults to the name. May contain HTML.
-        """
-        return self.name
-
-    def consume_items(self, offer, basket, affected_lines):
+if not is_model_registered('offer', 'Benefit'):
+    class Benefit(AbstractBenefit):
         pass
 
-    def is_satisfied(self, offer, basket):
-        """
-        Determines whether a given basket meets this condition.  This is
-        stubbed in this top-class object.  The subclassing proxies are
-        responsible for implementing it correctly.
-        """
-        return False
-
-    def is_partially_satisfied(self, offer, basket):
-        """
-        Determine if the basket partially meets the condition.  This is useful
-        for up-selling messages to entice customers to buy something more in
-        order to qualify for an offer.
-        """
-        return False
-
-    def get_upsell_message(self, offer, basket):
-        return None
-
-    def can_apply_condition(self, line):
-        """
-        Determines whether the condition can be applied to a given basket line
-        """
-        if not line.stockrecord_id:
-            return False
-        product = line.product
-        return (self.range.contains_product(product)
-                and product.get_is_discountable())
-
-    def get_applicable_lines(self, offer, basket, most_expensive_first=True):
-        """
-        Return line data for the lines that can be consumed by this condition
-        """
-        line_tuples = []
-        for line in basket.all_lines():
-            if not self.can_apply_condition(line):
-                continue
-
-            price = unit_price(offer, line)
-            if not price:
-                continue
-            line_tuples.append((price, line))
-        key = operator.itemgetter(0)
-        if most_expensive_first:
-            return sorted(line_tuples, reverse=True, key=key)
-        return sorted(line_tuples, key=key)
+    __all__.append('Benefit')
 
 
-@python_2_unicode_compatible
-class Benefit(models.Model):
-    range = models.ForeignKey(
-        'offer.Range', null=True, blank=True, verbose_name=_("Range"))
+if not is_model_registered('offer', 'Condition'):
+    class Condition(AbstractCondition):
+        pass
 
-    # Benefit types
-    PERCENTAGE, FIXED, MULTIBUY, FIXED_PRICE = (
-        "Percentage", "Absolute", "Multibuy", "Fixed price")
-    SHIPPING_PERCENTAGE, SHIPPING_ABSOLUTE, SHIPPING_FIXED_PRICE = (
-        'Shipping percentage', 'Shipping absolute', 'Shipping fixed price')
-    TYPE_CHOICES = (
-        (PERCENTAGE, _("Discount is a percentage off of the product's value")),
-        (FIXED, _("Discount is a fixed amount off of the product's value")),
-        (MULTIBUY, _("Discount is to give the cheapest product for free")),
-        (FIXED_PRICE,
-         _("Get the products that meet the condition for a fixed price")),
-        (SHIPPING_ABSOLUTE,
-         _("Discount is a fixed amount of the shipping cost")),
-        (SHIPPING_FIXED_PRICE, _("Get shipping for a fixed price")),
-        (SHIPPING_PERCENTAGE, _("Discount is a percentage off of the shipping"
-                                " cost")),
-    )
-    type = models.CharField(
-        _("Type"), max_length=128, choices=TYPE_CHOICES, blank=True)
-
-    # The value to use with the designated type.  This can be either an integer
-    # (eg for multibuy) or a decimal (eg an amount) which is slightly
-    # confusing.
-    value = fields.PositiveDecimalField(
-        _("Value"), decimal_places=2, max_digits=12, null=True, blank=True)
-
-    # If this is not set, then there is no upper limit on how many products
-    # can be discounted by this benefit.
-    max_affected_items = models.PositiveIntegerField(
-        _("Max Affected Items"), blank=True, null=True,
-        help_text=_("Set this to prevent the discount consuming all items "
-                    "within the range that are in the basket."))
-
-    # A custom benefit class can be used instead.  This means the
-    # type/value/max_affected_items fields should all be None.
-    proxy_class = fields.NullCharField(
-        _("Custom class"), max_length=255, unique=True, default=None)
-
-    class Meta:
-        app_label = 'offer'
-        verbose_name = _("Benefit")
-        verbose_name_plural = _("Benefits")
-
-    def proxy(self):
-        klassmap = {
-            self.PERCENTAGE: PercentageDiscountBenefit,
-            self.FIXED: AbsoluteDiscountBenefit,
-            self.MULTIBUY: MultibuyDiscountBenefit,
-            self.FIXED_PRICE: FixedPriceBenefit,
-            self.SHIPPING_ABSOLUTE: ShippingAbsoluteDiscountBenefit,
-            self.SHIPPING_FIXED_PRICE: ShippingFixedPriceBenefit,
-            self.SHIPPING_PERCENTAGE: ShippingPercentageDiscountBenefit}
-        # Short-circuit logic if current class is already a proxy class.
-        if self.__class__ in klassmap.values():
-            return self
-
-        field_dict = dict(self.__dict__)
-        for field in list(field_dict.keys()):
-            if field.startswith('_'):
-                del field_dict[field]
-
-        if self.proxy_class:
-            klass = load_proxy(self.proxy_class)
-            # Short-circuit again.
-            if self.__class__ == klass:
-                return self
-            return klass(**field_dict)
-
-        if self.type in klassmap:
-            return klassmap[self.type](**field_dict)
-        raise RuntimeError("Unrecognised benefit type (%s)" % self.type)
-
-    def __str__(self):
-        return self.name
-
-    @property
-    def name(self):
-        """
-        A plaintext description of the benefit. Every proxy class has to
-        implement it.
-
-        This is used in the dropdowns within the offer dashboard.
-        """
-        return self.proxy().name
-
-    @property
-    def description(self):
-        """
-        A description of the benefit.
-        Defaults to the name. May contain HTML.
-        """
-        return self.name
-
-    def apply(self, basket, condition, offer):
-        return ZERO_DISCOUNT
-
-    def apply_deferred(self, basket, order, application):
-        return None
-
-    def clean(self):
-        if not self.type:
-            return
-        method_name = 'clean_%s' % self.type.lower().replace(' ', '_')
-        if hasattr(self, method_name):
-            getattr(self, method_name)()
-
-    def clean_multibuy(self):
-        if not self.range:
-            raise ValidationError(
-                _("Multibuy benefits require a product range"))
-        if self.value:
-            raise ValidationError(
-                _("Multibuy benefits don't require a value"))
-        if self.max_affected_items:
-            raise ValidationError(
-                _("Multibuy benefits don't require a 'max affected items' "
-                  "attribute"))
-
-    def clean_percentage(self):
-        if not self.range:
-            raise ValidationError(
-                _("Percentage benefits require a product range"))
-        if self.value > 100:
-            raise ValidationError(
-                _("Percentage discount cannot be greater than 100"))
-
-    def clean_shipping_absolute(self):
-        if not self.value:
-            raise ValidationError(
-                _("A discount value is required"))
-        if self.range:
-            raise ValidationError(
-                _("No range should be selected as this benefit does not "
-                  "apply to products"))
-        if self.max_affected_items:
-            raise ValidationError(
-                _("Shipping discounts don't require a 'max affected items' "
-                  "attribute"))
-
-    def clean_shipping_percentage(self):
-        if self.value > 100:
-            raise ValidationError(
-                _("Percentage discount cannot be greater than 100"))
-        if self.range:
-            raise ValidationError(
-                _("No range should be selected as this benefit does not "
-                  "apply to products"))
-        if self.max_affected_items:
-            raise ValidationError(
-                _("Shipping discounts don't require a 'max affected items' "
-                  "attribute"))
-
-    def clean_shipping_fixed_price(self):
-        if self.range:
-            raise ValidationError(
-                _("No range should be selected as this benefit does not "
-                  "apply to products"))
-        if self.max_affected_items:
-            raise ValidationError(
-                _("Shipping discounts don't require a 'max affected items' "
-                  "attribute"))
-
-    def clean_fixed_price(self):
-        if self.range:
-            raise ValidationError(
-                _("No range should be selected as the condition range will "
-                  "be used instead."))
-
-    def clean_absolute(self):
-        if not self.range:
-            raise ValidationError(
-                _("Fixed discount benefits require a product range"))
-        if not self.value:
-            raise ValidationError(
-                _("Fixed discount benefits require a value"))
-
-    def round(self, amount):
-        """
-        Apply rounding to discount amount
-        """
-        if hasattr(settings, 'OSCAR_OFFER_ROUNDING_FUNCTION'):
-            return settings.OSCAR_OFFER_ROUNDING_FUNCTION(amount)
-        return amount.quantize(D('.01'), ROUND_DOWN)
-
-    def _effective_max_affected_items(self):
-        """
-        Return the maximum number of items that can have a discount applied
-        during the application of this benefit
-        """
-        return self.max_affected_items if self.max_affected_items else 10000
-
-    def can_apply_benefit(self, line):
-        """
-        Determines whether the benefit can be applied to a given basket line
-        """
-        return line.stockrecord and line.product.is_discountable
-
-    def get_applicable_lines(self, offer, basket, range=None):
-        """
-        Return the basket lines that are available to be discounted
-
-        :basket: The basket
-        :range: The range of products to use for filtering.  The fixed-price
-                benefit ignores its range and uses the condition range
-        """
-        if range is None:
-            range = self.range
-        line_tuples = []
-        for line in basket.all_lines():
-            product = line.product
-
-            if (not range.contains(product) or
-                    not self.can_apply_benefit(line)):
-                continue
-
-            price = unit_price(offer, line)
-            if not price:
-                # Avoid zero price products
-                continue
-            if line.quantity_without_discount == 0:
-                continue
-            line_tuples.append((price, line))
-
-        # We sort lines to be cheapest first to ensure consistent applications
-        return sorted(line_tuples, key=operator.itemgetter(0))
-
-    def shipping_discount(self, charge):
-        return D('0.00')
+    __all__.append('Condition')
 
 
 if not is_model_registered('offer', 'Range'):
@@ -441,716 +48,21 @@ if not is_model_registered('offer', 'RangeProduct'):
     __all__.append('RangeProduct')
 
 
-# ==========
-# Conditions
-# ==========
+if not is_model_registered('offer', 'RangeProductFileUpload'):
+    class RangeProductFileUpload(AbstractRangeProductFileUpload):
+        pass
 
+    __all__.append('RangeProductFileUpload')
 
-class CountCondition(Condition):
-    """
-    An offer condition dependent on the NUMBER of matching items from the
-    basket.
-    """
-    _description = _("Basket includes %(count)d item(s) from %(range)s")
 
-    @property
-    def name(self):
-        return self._description % {
-            'count': self.value,
-            'range': six.text_type(self.range).lower()}
+# Import the benefits and the conditions. Required after initializing the
+# parent models to allow overriding them
 
-    @property
-    def description(self):
-        return self._description % {
-            'count': self.value,
-            'range': range_anchor(self.range)}
+from oscar.apps.offer.benefits import *  # noqa
+from oscar.apps.offer.conditions import *  # noqa
 
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Count condition")
-        verbose_name_plural = _("Count conditions")
+from oscar.apps.offer.benefits import __all__ as benefit_classes  # noqa
+from oscar.apps.offer.conditions import __all__ as condition_classes  # noqa
 
-    def is_satisfied(self, offer, basket):
-        """
-        Determines whether a given basket meets this condition
-        """
-        num_matches = 0
-        for line in basket.all_lines():
-            if (self.can_apply_condition(line)
-                    and line.quantity_without_discount > 0):
-                num_matches += line.quantity_without_discount
-            if num_matches >= self.value:
-                return True
-        return False
-
-    def _get_num_matches(self, basket):
-        if hasattr(self, '_num_matches'):
-            return getattr(self, '_num_matches')
-        num_matches = 0
-        for line in basket.all_lines():
-            if (self.can_apply_condition(line)
-                    and line.quantity_without_discount > 0):
-                num_matches += line.quantity_without_discount
-        self._num_matches = num_matches
-        return num_matches
-
-    def is_partially_satisfied(self, offer, basket):
-        num_matches = self._get_num_matches(basket)
-        return 0 < num_matches < self.value
-
-    def get_upsell_message(self, offer, basket):
-        num_matches = self._get_num_matches(basket)
-        delta = self.value - num_matches
-        return ungettext('Buy %(delta)d more product from %(range)s',
-                         'Buy %(delta)d more products from %(range)s', delta) \
-            % {'delta': delta, 'range': self.range}
-
-    def consume_items(self, offer, basket, affected_lines):
-        """
-        Marks items within the basket lines as consumed so they
-        can't be reused in other offers.
-
-        :basket: The basket
-        :affected_lines: The lines that have been affected by the discount.
-                         This should be list of tuples (line, discount, qty)
-        """
-        # We need to count how many items have already been consumed as part of
-        # applying the benefit, so we don't consume too many items.
-        num_consumed = 0
-        for line, __, quantity in affected_lines:
-            num_consumed += quantity
-        to_consume = max(0, self.value - num_consumed)
-        if to_consume == 0:
-            return
-
-        for __, line in self.get_applicable_lines(offer, basket,
-                                                  most_expensive_first=True):
-            quantity_to_consume = min(line.quantity_without_discount,
-                                      to_consume)
-            line.consume(quantity_to_consume)
-            to_consume -= quantity_to_consume
-            if to_consume == 0:
-                break
-
-
-class CoverageCondition(Condition):
-    """
-    An offer condition dependent on the number of DISTINCT matching items from
-    the basket.
-    """
-    _description = _("Basket includes %(count)d distinct item(s) from"
-                     " %(range)s")
-
-    @property
-    def name(self):
-        return self._description % {
-            'count': self.value,
-            'range': six.text_type(self.range).lower()}
-
-    @property
-    def description(self):
-        return self._description % {
-            'count': self.value,
-            'range': range_anchor(self.range)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Coverage Condition")
-        verbose_name_plural = _("Coverage Conditions")
-
-    def is_satisfied(self, offer, basket):
-        """
-        Determines whether a given basket meets this condition
-        """
-        covered_ids = []
-        for line in basket.all_lines():
-            if not line.is_available_for_discount:
-                continue
-            product = line.product
-            if (self.can_apply_condition(line) and product.id not in
-                    covered_ids):
-                covered_ids.append(product.id)
-            if len(covered_ids) >= self.value:
-                return True
-        return False
-
-    def _get_num_covered_products(self, basket):
-        covered_ids = []
-        for line in basket.all_lines():
-            if not line.is_available_for_discount:
-                continue
-            product = line.product
-            if (self.can_apply_condition(line) and product.id not in
-                    covered_ids):
-                covered_ids.append(product.id)
-        return len(covered_ids)
-
-    def get_upsell_message(self, offer, basket):
-        delta = self.value - self._get_num_covered_products(basket)
-        return ungettext('Buy %(delta)d more product from %(range)s',
-                         'Buy %(delta)d more products from %(range)s', delta) \
-            % {'delta': delta, 'range': self.range}
-
-    def is_partially_satisfied(self, offer, basket):
-        return 0 < self._get_num_covered_products(basket) < self.value
-
-    def consume_items(self, offer, basket, affected_lines):
-        """
-        Marks items within the basket lines as consumed so they
-        can't be reused in other offers.
-        """
-        # Determine products that have already been consumed by applying the
-        # benefit
-        consumed_products = []
-        for line, __, quantity in affected_lines:
-            consumed_products.append(line.product)
-
-        to_consume = max(0, self.value - len(consumed_products))
-        if to_consume == 0:
-            return
-
-        for line in basket.all_lines():
-            product = line.product
-            if not self.can_apply_condition(line):
-                continue
-            if product in consumed_products:
-                continue
-            if not line.is_available_for_discount:
-                continue
-            # Only consume a quantity of 1 from each line
-            line.consume(1)
-            consumed_products.append(product)
-            to_consume -= 1
-            if to_consume == 0:
-                break
-
-    def get_value_of_satisfying_items(self, offer, basket):
-        covered_ids = []
-        value = D('0.00')
-        for line in basket.all_lines():
-            if (self.can_apply_condition(line) and line.product.id not in
-                    covered_ids):
-                covered_ids.append(line.product.id)
-                value += unit_price(offer, line)
-            if len(covered_ids) >= self.value:
-                return value
-        return value
-
-
-class ValueCondition(Condition):
-    """
-    An offer condition dependent on the VALUE of matching items from the
-    basket.
-    """
-    _description = _("Basket includes %(amount)s from %(range)s")
-
-    @property
-    def name(self):
-        return self._description % {
-            'amount': currency(self.value),
-            'range': six.text_type(self.range).lower()}
-
-    @property
-    def description(self):
-        return self._description % {
-            'amount': currency(self.value),
-            'range': range_anchor(self.range)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Value condition")
-        verbose_name_plural = _("Value conditions")
-
-    def is_satisfied(self, offer, basket):
-        """
-        Determine whether a given basket meets this condition
-        """
-        value_of_matches = D('0.00')
-        for line in basket.all_lines():
-            if (self.can_apply_condition(line) and
-                    line.quantity_without_discount > 0):
-                price = unit_price(offer, line)
-                value_of_matches += price * int(line.quantity_without_discount)
-            if value_of_matches >= self.value:
-                return True
-        return False
-
-    def _get_value_of_matches(self, offer, basket):
-        if hasattr(self, '_value_of_matches'):
-            return getattr(self, '_value_of_matches')
-        value_of_matches = D('0.00')
-        for line in basket.all_lines():
-            if (self.can_apply_condition(line) and
-                    line.quantity_without_discount > 0):
-                price = unit_price(offer, line)
-                value_of_matches += price * int(line.quantity_without_discount)
-        self._value_of_matches = value_of_matches
-        return value_of_matches
-
-    def is_partially_satisfied(self, offer, basket):
-        value_of_matches = self._get_value_of_matches(offer, basket)
-        return D('0.00') < value_of_matches < self.value
-
-    def get_upsell_message(self, offer, basket):
-        value_of_matches = self._get_value_of_matches(offer, basket)
-        return _('Spend %(value)s more from %(range)s') % {
-            'value': currency(self.value - value_of_matches),
-            'range': self.range}
-
-    def consume_items(self, offer, basket, affected_lines):
-        """
-        Marks items within the basket lines as consumed so they
-        can't be reused in other offers.
-
-        We allow lines to be passed in as sometimes we want them sorted
-        in a specific order.
-        """
-        # Determine value of items already consumed as part of discount
-        value_consumed = D('0.00')
-        for line, __, qty in affected_lines:
-            price = unit_price(offer, line)
-            value_consumed += price * qty
-
-        to_consume = max(0, self.value - value_consumed)
-        if to_consume == 0:
-            return
-
-        for price, line in self.get_applicable_lines(
-                offer, basket, most_expensive_first=True):
-            quantity_to_consume = min(
-                line.quantity_without_discount,
-                (to_consume / price).quantize(D(1), ROUND_UP))
-            line.consume(quantity_to_consume)
-            to_consume -= price * quantity_to_consume
-            if to_consume <= 0:
-                break
-
-
-# ========
-# Benefits
-# ========
-
-
-class PercentageDiscountBenefit(Benefit):
-    """
-    An offer benefit that gives a percentage discount
-    """
-    _description = _("%(value)s%% discount on %(range)s")
-
-    @property
-    def name(self):
-        return self._description % {
-            'value': self.value,
-            'range': self.range.name}
-
-    @property
-    def description(self):
-        return self._description % {
-            'value': self.value,
-            'range': range_anchor(self.range)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Percentage discount benefit")
-        verbose_name_plural = _("Percentage discount benefits")
-
-    def apply(self, basket, condition, offer, discount_percent=None,
-              max_total_discount=None):
-        if discount_percent is None:
-            discount_percent = self.value
-
-        discount_amount_available = max_total_discount
-
-        line_tuples = self.get_applicable_lines(offer, basket)
-
-        discount = D('0.00')
-        affected_items = 0
-        max_affected_items = self._effective_max_affected_items()
-        affected_lines = []
-        for price, line in line_tuples:
-            if affected_items >= max_affected_items:
-                break
-            if discount_amount_available == 0:
-                break
-
-            quantity_affected = min(line.quantity_without_discount,
-                                    max_affected_items - affected_items)
-            line_discount = self.round(discount_percent / D('100.0') * price
-                                       * int(quantity_affected))
-
-            if discount_amount_available is not None:
-                line_discount = min(line_discount, discount_amount_available)
-                discount_amount_available -= line_discount
-
-            apply_discount(line, line_discount, quantity_affected)
-
-            affected_lines.append((line, line_discount, quantity_affected))
-            affected_items += quantity_affected
-            discount += line_discount
-
-        if discount > 0:
-            condition.consume_items(offer, basket, affected_lines)
-        return BasketDiscount(discount)
-
-
-class AbsoluteDiscountBenefit(Benefit):
-    """
-    An offer benefit that gives an absolute discount
-    """
-    _description = _("%(value)s discount on %(range)s")
-
-    @property
-    def name(self):
-        return self._description % {
-            'value': currency(self.value),
-            'range': self.range.name.lower()}
-
-    @property
-    def description(self):
-        return self._description % {
-            'value': currency(self.value),
-            'range': range_anchor(self.range)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Absolute discount benefit")
-        verbose_name_plural = _("Absolute discount benefits")
-
-    def apply(self, basket, condition, offer, discount_amount=None,
-              max_total_discount=None):
-        if discount_amount is None:
-            discount_amount = self.value
-
-        # Fetch basket lines that are in the range and available to be used in
-        # an offer.
-        line_tuples = self.get_applicable_lines(offer, basket)
-
-        # Determine which lines can have the discount applied to them
-        max_affected_items = self._effective_max_affected_items()
-        num_affected_items = 0
-        affected_items_total = D('0.00')
-        lines_to_discount = []
-        for price, line in line_tuples:
-            if num_affected_items >= max_affected_items:
-                break
-            qty = min(line.quantity_without_discount,
-                      max_affected_items - num_affected_items)
-            lines_to_discount.append((line, price, qty))
-            num_affected_items += qty
-            affected_items_total += qty * price
-
-        # Ensure we don't try to apply a discount larger than the total of the
-        # matching items.
-        discount = min(discount_amount, affected_items_total)
-        if max_total_discount is not None:
-            discount = min(discount, max_total_discount)
-
-        if discount == 0:
-            return ZERO_DISCOUNT
-
-        # Apply discount equally amongst them
-        affected_lines = []
-        applied_discount = D('0.00')
-        for i, (line, price, qty) in enumerate(lines_to_discount):
-            if i == len(lines_to_discount) - 1:
-                # If last line, then take the delta as the discount to ensure
-                # the total discount is correct and doesn't mismatch due to
-                # rounding.
-                line_discount = discount - applied_discount
-            else:
-                # Calculate a weighted discount for the line
-                line_discount = self.round(
-                    ((price * qty) / affected_items_total) * discount)
-            apply_discount(line, line_discount, qty)
-            affected_lines.append((line, line_discount, qty))
-            applied_discount += line_discount
-
-        condition.consume_items(offer, basket, affected_lines)
-
-        return BasketDiscount(discount)
-
-
-class FixedPriceBenefit(Benefit):
-    """
-    An offer benefit that gives the items in the condition for a
-    fixed price.  This is useful for "bundle" offers.
-
-    Note that we ignore the benefit range here and only give a fixed price
-    for the products in the condition range.  The condition cannot be a value
-    condition.
-
-    We also ignore the max_affected_items setting.
-    """
-    _description = _("The products that meet the condition are sold "
-                     "for %(amount)s")
-
-    @property
-    def name(self):
-        return self._description % {
-            'amount': currency(self.value)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Fixed price benefit")
-        verbose_name_plural = _("Fixed price benefits")
-
-    def apply(self, basket, condition, offer):  # noqa (too complex (10))
-        if isinstance(condition, ValueCondition):
-            return ZERO_DISCOUNT
-
-        # Fetch basket lines that are in the range and available to be used in
-        # an offer.
-        line_tuples = self.get_applicable_lines(offer, basket,
-                                                range=condition.range)
-        if not line_tuples:
-            return ZERO_DISCOUNT
-
-        # Determine the lines to consume
-        num_permitted = int(condition.value)
-        num_affected = 0
-        value_affected = D('0.00')
-        covered_lines = []
-        for price, line in line_tuples:
-            if isinstance(condition, CoverageCondition):
-                quantity_affected = 1
-            else:
-                quantity_affected = min(
-                    line.quantity_without_discount,
-                    num_permitted - num_affected)
-            num_affected += quantity_affected
-            value_affected += quantity_affected * price
-            covered_lines.append((price, line, quantity_affected))
-            if num_affected >= num_permitted:
-                break
-        discount = max(value_affected - self.value, D('0.00'))
-        if not discount:
-            return ZERO_DISCOUNT
-
-        # Apply discount to the affected lines
-        discount_applied = D('0.00')
-        last_line = covered_lines[-1][1]
-        for price, line, quantity in covered_lines:
-            if line == last_line:
-                # If last line, we just take the difference to ensure that
-                # rounding doesn't lead to an off-by-one error
-                line_discount = discount - discount_applied
-            else:
-                line_discount = self.round(
-                    discount * (price * quantity) / value_affected)
-            apply_discount(line, line_discount, quantity)
-            discount_applied += line_discount
-        return BasketDiscount(discount)
-
-
-class MultibuyDiscountBenefit(Benefit):
-    _description = _("Cheapest product from %(range)s is free")
-
-    @property
-    def name(self):
-        return self._description % {
-            'range': self.range.name.lower()}
-
-    @property
-    def description(self):
-        return self._description % {
-            'range': range_anchor(self.range)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Multibuy discount benefit")
-        verbose_name_plural = _("Multibuy discount benefits")
-
-    def apply(self, basket, condition, offer):
-        line_tuples = self.get_applicable_lines(offer, basket)
-        if not line_tuples:
-            return ZERO_DISCOUNT
-
-        # Cheapest line gives free product
-        discount, line = line_tuples[0]
-        apply_discount(line, discount, 1)
-
-        affected_lines = [(line, discount, 1)]
-        condition.consume_items(offer, basket, affected_lines)
-
-        return BasketDiscount(discount)
-
-
-# =================
-# Shipping benefits
-# =================
-
-
-class ShippingBenefit(Benefit):
-
-    def apply(self, basket, condition, offer):
-        condition.consume_items(offer, basket, affected_lines=())
-        return SHIPPING_DISCOUNT
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-
-
-class ShippingAbsoluteDiscountBenefit(ShippingBenefit):
-    _description = _("%(amount)s off shipping cost")
-
-    @property
-    def name(self):
-        return self._description % {
-            'amount': currency(self.value)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Shipping absolute discount benefit")
-        verbose_name_plural = _("Shipping absolute discount benefits")
-
-    def shipping_discount(self, charge):
-        return min(charge, self.value)
-
-
-class ShippingFixedPriceBenefit(ShippingBenefit):
-    _description = _("Get shipping for %(amount)s")
-
-    @property
-    def name(self):
-        return self._description % {
-            'amount': currency(self.value)}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Fixed price shipping benefit")
-        verbose_name_plural = _("Fixed price shipping benefits")
-
-    def shipping_discount(self, charge):
-        if charge < self.value:
-            return D('0.00')
-        return charge - self.value
-
-
-class ShippingPercentageDiscountBenefit(ShippingBenefit):
-    _description = _("%(value)s%% off of shipping cost")
-
-    @property
-    def name(self):
-        return self._description % {
-            'value': self.value}
-
-    class Meta:
-        app_label = 'offer'
-        proxy = True
-        verbose_name = _("Shipping percentage discount benefit")
-        verbose_name_plural = _("Shipping percentage discount benefits")
-
-    def shipping_discount(self, charge):
-        discount = charge * self.value / D('100.0')
-        return discount.quantize(D('0.01'))
-
-
-class RangeProductFileUpload(models.Model):
-    range = models.ForeignKey('offer.Range', related_name='file_uploads',
-                              verbose_name=_("Range"))
-    filepath = models.CharField(_("File Path"), max_length=255)
-    size = models.PositiveIntegerField(_("Size"))
-    uploaded_by = models.ForeignKey(AUTH_USER_MODEL,
-                                    verbose_name=_("Uploaded By"))
-    date_uploaded = models.DateTimeField(_("Date Uploaded"), auto_now_add=True)
-
-    PENDING, FAILED, PROCESSED = 'Pending', 'Failed', 'Processed'
-    choices = (
-        (PENDING, PENDING),
-        (FAILED, FAILED),
-        (PROCESSED, PROCESSED),
-    )
-    status = models.CharField(_("Status"), max_length=32, choices=choices,
-                              default=PENDING)
-    error_message = models.CharField(_("Error Message"), max_length=255,
-                                     blank=True)
-
-    # Post-processing audit fields
-    date_processed = models.DateTimeField(_("Date Processed"), null=True)
-    num_new_skus = models.PositiveIntegerField(_("Number of New SKUs"),
-                                               null=True)
-    num_unknown_skus = models.PositiveIntegerField(_("Number of Unknown SKUs"),
-                                                   null=True)
-    num_duplicate_skus = models.PositiveIntegerField(
-        _("Number of Duplicate SKUs"), null=True)
-
-    class Meta:
-        app_label = 'offer'
-        ordering = ('-date_uploaded',)
-        verbose_name = _("Range Product Uploaded File")
-        verbose_name_plural = _("Range Product Uploaded Files")
-
-    @property
-    def filename(self):
-        return os.path.basename(self.filepath)
-
-    def mark_as_failed(self, message=None):
-        self.date_processed = now()
-        self.error_message = message
-        self.status = self.FAILED
-        self.save()
-
-    def mark_as_processed(self, num_new, num_unknown, num_duplicate):
-        self.status = self.PROCESSED
-        self.date_processed = now()
-        self.num_new_skus = num_new
-        self.num_unknown_skus = num_unknown
-        self.num_duplicate_skus = num_duplicate
-        self.save()
-
-    def was_processing_successful(self):
-        return self.status == self.PROCESSED
-
-    def process(self):
-        """
-        Process the file upload and add products to the range
-        """
-        all_ids = set(self.extract_ids())
-        products = self.range.included_products.all()
-        existing_skus = products.values_list(
-            'stockrecords__partner_sku', flat=True)
-        existing_skus = set(filter(bool, existing_skus))
-        existing_upcs = products.values_list('upc', flat=True)
-        existing_upcs = set(filter(bool, existing_upcs))
-        existing_ids = existing_skus.union(existing_upcs)
-        new_ids = all_ids - existing_ids
-
-        Product = models.get_model('catalogue', 'Product')
-        products = Product._default_manager.filter(
-            models.Q(stockrecords__partner_sku__in=new_ids) |
-            models.Q(upc__in=new_ids))
-        for product in products:
-            self.range.add_product(product)
-
-        # Processing stats
-        found_skus = products.values_list(
-            'stockrecords__partner_sku', flat=True)
-        found_skus = set(filter(bool, found_skus))
-        found_upcs = set(filter(bool, products.values_list('upc', flat=True)))
-        found_ids = found_skus.union(found_upcs)
-        missing_ids = new_ids - found_ids
-        dupes = set(all_ids).intersection(existing_ids)
-
-        self.mark_as_processed(products.count(), len(missing_ids), len(dupes))
-
-    def extract_ids(self):
-        """
-        Extract all SKU- or UPC-like strings from the file
-        """
-        for line in open(self.filepath, 'r'):
-            for id in re.split('[^\w:\.-]', line):
-                if id:
-                    yield id
-
-    def delete_file(self):
-        os.unlink(self.filepath)
+__all__.extend(benefit_classes)
+__all__.extend(condition_classes)
